project 35
Add a “waiting” indicator after the Submit button is pressed. The screen should grey out
the page so that user does not re-click submit, and a spinning animation should be
displayed. Following link may be useful.
https://www.hungred.com/how-to/tutorial-simple-screen-grey-effect-jquery/
Use the provided animated gif and display it in the center of the screen.