project 38 3-1
Once the user clicks “SUBMIT” in the upload window, return the user to the website and
execute the upload to S3 logic without the user having to wait. A background process is
initiated to take care of the upload logic and then start the tiling process on the EC2
server.