project 46
1. After an asset is deleted (see project 43), re-direct the user to
https://gw3.construkted.com/profile/?active_tab=posts

3. Front end submission form: Before accepting a file input, test to make sure the node
server is running. If it’s not running throw an error asking the user to contact
“support@construkted.com”
The problem is that if the tiling node server is not running, then uploaded files may stay
on the wordpress server, and never get moved to S3.
Wordpress server is not for file storage.